15.08 Update+
- Fix Play services battery drain on arm
- Fixed remove-faceunlock from updater script
- Show some love for who doesn't have realpath
- Export md5 as a file too
- Simplify create checks (@stucki 6a05464df7ae80784fc8fe8ff2de65152d7bb67b)
- Move target zip out of arch directory  (@stucki bf57998702c634f560535b586706931f56a3ed31)
- Clean Up after the build (@stucki f6660501afa1eaefc4d3cf1a043e5c73bdab398d)
- CleanUp final output

+++ build/gapps.sh
@@ -25,26 +25,28 @@ function printdone(){
   echo "$(tput setaf 2)$1$(tput sgr 0)"
 }
 
+function clean(){
+    echo "Cleaning up..."
+    rm -r $OUT/$GARCH
+    rm /tmp/$BUILDZIP
+    return $?
+}
+
+function Gfailed(){
+    printerr "Build failed, check $GLOG"
+    exit 1
+}
+
 function create(){
-    if [ -f $GLOG ]; then
-        rm -f $GLOG
-    fi
+    test -f $GLOG && rm -f $GLOG
     echo "Starting GApps compilation" > $GLOG
     echo "ARCH= $GARCH" >> $GLOG
     echo "OS= $(uname -s -r)" >> $GLOG
     echo "NAME= $(whoami) at $(uname -n)" >> $GLOG
     PREBUILT=$TOP/prebuilt/gapps/$GARCH
-    if [ -d $OUT/$GARCH ]; then
-        echo "Previous build found for $GARCH!" >> $GLOG
-    else
-        echo "No previous build found for $GARCH!" >> $GLOG
-        if [ -d $OUT ]; then
-            echo "OUT directory detected at: $OUT" >> $GLOG
-        else
-            mkdir $OUT
-        fi
-        mkdir $OUT/$GARCH && echo "Created build directories" >> $GLOG
-    fi
+    test -d $OUT || mkdir $OUT;
+    test -d $OUT/$GARCH || mkdir -p $OUT/$GARCH
+    echo "Build directories are now ready" >> $GLOG
     echo "Getting prebuilts..."
     echo "Copying stuffs" >> $GLOG
     cp -r $PREBUILT $OUT/$GARCH >> $GLOG
@@ -53,7 +55,7 @@ function create(){
 }
 
 function zipit(){
-    BUILDZIP=gapps-$ANDROIDV-$DATE.zip
+    BUILDZIP=gapps-$ANDROIDV-$GARCH-$DATE.zip
     echo "Importing installation scripts..."
     cp -r $METAINF $OUT/$GARCH/META-INF && echo "Meta copied" >> $GLOG
     echo "Creating package..."
@@ -63,7 +65,7 @@ function zipit(){
     cd $TOP
     if [ -f /tmp/$BUILDZIP ]; then
         echo "Signing zip..."
-        java -Xmx2048m -jar $TOP/build/sign/signapk.jar -w $TOP/build/sign/testkey.x509.pem $TOP/build/sign/testkey.pk8 /tmp/$BUILDZIP $OUT/$GARCH/$BUILDZIP >> $GLOG
+        java -Xmx2048m -jar $TOP/build/sign/signapk.jar -w $TOP/build/sign/testkey.x509.pem $TOP/build/sign/testkey.pk8 /tmp/$BUILDZIP $OUT/$BUILDZIP >> $GLOG
     else
         printerr "Couldn't zip files!"
         echo "Couldn't find unsigned zip file, aborting" >> $GLOG
@@ -75,7 +77,9 @@ function getmd5(){
     if [ -x $(which md5sum) ]; then
         echo "md5sum is installed, getting md5..." >> $GLOG
         echo "Getting md5sum..."
-        GMD5=$(md5sum $OUT/$GARCH/$BUILDZIP)
+        GMD5=$(md5sum $OUT/$BUILDZIP)
+        echo -e "$GMD5" > $OUT/gapps-$ANDROIDV-$GARCH-$DATE.md5
+        echo "md5 exported at $OUT/gapps-$ANDROIDV-$GARCH-$DATE.md5"
         return 0
     else
         echo "md5sum is not installed, aborting" >> $GLOG
@@ -89,6 +93,12 @@ function getmd5(){
 GARCH=$1
 create
 LASTRETURN=$?
+if [ -x $(which realpath) ]; then
+    echo "Realpath found!" >> $GLOG
+else
+    TOP=$(cd . && pwd) # some os X love
+    echo "No realpath found!" >> $GLOG
+fi
 if [ "$LASTRETURN" == 0 ]; then
     zipit
     LASTRETURN=$?
@@ -96,19 +106,21 @@ if [ "$LASTRETURN" == 0 ]; then
         getmd5
         LASTRETURN=$?
         if [ "$LASTRETURN" == 0 ]; then
-            echo "Done!" >> $GLOG
-            printdone "Build completed: $OUT/$GARCH/$BUILDZIP"
-            printdone "            md5: $GMD5"
-            exit 0
+            clean
+            LASTRETURN=$?
+            if [ "$LASTRETURN" == 0 ]; then
+                echo "Done!" >> $GLOG
+                printdone "Build completed: $GMD5"
+                exit 0
+            else
+                Gfailed
+            fi
         else
-            printerr "Build failed, check $GLOG"
-            exit 1
+            Gfailed
         fi
     else
-        printerr "Build failed, check $GLOG"
-        exit 1
+        Gfailed
     fi
 else
-    printerr "Build failed, check $GLOG"
-    exit 1
+    Gfailed
 fi
